91_перенос методов в базовый фрагмент
