+++ app/src/main/java/com/davay/android/feature/sessionconnection/presentation/SessionConnectionBottomSheetFragment.kt
@@ -3,17 +3,10 @@ package com.davay.android.feature.sessionconnection.presentation
 import android.app.Dialog
 import android.os.Bundle
 import android.view.View
-import android.view.ViewGroup
 import android.view.ViewTreeObserver
-import android.view.WindowManager
 import android.view.inputmethod.EditorInfo
 import android.view.inputmethod.InputMethodManager
 import androidx.core.content.ContextCompat
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsAnimationCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.core.view.updateLayoutParams
 import androidx.core.widget.doAfterTextChanged
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
@@ -39,77 +32,14 @@ class SessionConnectionBottomSheetFragment :
         .appComponent(AppComponentHolder.getComponent())
         .build()
 
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val thisDdailog = super.onCreateDialog(savedInstanceState)
-        thisDdailog.window?.also {
-            it.setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_STATE_ALWAYS_VISIBLE)
-            WindowCompat.setDecorFitsSystemWindows(it, false)
-        }
-        return thisDdailog
-    }
+    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
+        makeDialogWithKeyboard(savedInstanceState)
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initViews(view)
         subscribe()
-        animateKeyboard()
-    }
-
-    private fun animateKeyboard() {
-        ViewCompat.setOnApplyWindowInsetsListener(requireActivity().window.decorView) { _, windowInsets ->
-            try {
-                var insetsIme = windowInsets.getInsets(WindowInsetsCompat.Type.ime())
-                val insetsNav = windowInsets.getInsets(WindowInsetsCompat.Type.navigationBars())
-                if (windowInsets.isVisible(WindowInsetsCompat.Type.ime())) {
-                    binding.btnEnter.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-                        bottomMargin = insetsIme.bottom - insetsNav.bottom
-                    }
-                } else {
-                    binding.btnEnter.updateLayoutParams<ViewGroup.MarginLayoutParams> {
-                        bottomMargin = 0
-                    }
-                }
-            } catch (e: NullPointerException) {
-                // ничего не делаем
-                e.printStackTrace()
-            }
-            windowInsets
-        }
-
-        // api 30+
-        ViewCompat.setWindowInsetsAnimationCallback(
-            requireView().rootView,
-            object : WindowInsetsAnimationCompat.Callback(DISPATCH_MODE_STOP) {
-                var startBottom = 0f
-                var endBottom = 0f
-
-                override fun onPrepare(
-                    animation: WindowInsetsAnimationCompat
-                ) {
-                    startBottom = requireView().rootView.bottom.toFloat()
-                }
-
-                override fun onStart(
-                    animation: WindowInsetsAnimationCompat,
-                    bounds: WindowInsetsAnimationCompat.BoundsCompat
-                ): WindowInsetsAnimationCompat.BoundsCompat {
-                    endBottom = binding.root.bottom.toFloat()
-                    return bounds
-                }
-
-                override fun onProgress(
-                    insets: WindowInsetsCompat,
-                    runningAnimations: MutableList<WindowInsetsAnimationCompat>
-                ): WindowInsetsCompat {
-                    val imeAnimation = runningAnimations.find {
-                        it.typeMask and WindowInsetsCompat.Type.ime() != 0
-                    } ?: return insets
-                    binding.btnEnter.translationY =
-                        (startBottom - endBottom) * (1 - imeAnimation.interpolatedFraction)
-                    return insets
-                }
-            }
-        )
+        moveBottomView(binding.btnEnter)
     }
 
     private fun initViews(view: View) {
